+++ info/templates/about.html
@@ -5,7 +5,7 @@
 {% block content %}
 {% load static %}
 <video height="480" 
-       src="{% static 'pharmacy/videos/HealthHub_video.mp4' %}"
+       src="{% static 'info/videos/HealthHub_video.mp4' %}"
        autoplay controls preload="none" muted>
   Video cannot be displayed
 </video>
