+++ info/templates/index.html
@@ -8,10 +8,10 @@
 <a href="https://sinupret.com/">
     <figure>
         {% load static %}
-        <img src="{% static 'pharmacy/images/ad/ad_wide.jpg' %}" alt="Advertisement" width="1000"
+        <img src="{% static 'info/images/ad/ad_wide.jpg' %}" alt="Advertisement" width="1000"
 
-            srcset="{% static 'pharmacy/images/ad/ad_wide.jpg' %} 1920w,
-                    {% static 'pharmacy/images/ad/ad_short.jpg' %} 1200w"
+            srcset="{% static 'info/images/ad/ad_wide.jpg' %} 1920w,
+                    {% static 'info/images/ad/ad_short.jpg' %} 1200w"
         >
         <figcaption>Advertisement</figcaption>
     </figure>
@@ -29,19 +29,19 @@
 
 <h2>Out partners:</h2>
 <a href="https://www.jnj.com/">
-    <img src="{% static 'pharmacy/images/partners/jnj.svg' %}" alt="Jhonson and Jhonson" height="100">
+    <img src="{% static 'info/images/partners/jnj.svg' %}" alt="Jhonson and Jhonson" height="100">
 </a>
 <a href="https://www.gsk.com/">
-    <img src="{% static 'pharmacy/images/partners/gsk.png' %}" alt="GSK" height="100">
+    <img src="{% static 'info/images/partners/gsk.png' %}" alt="GSK" height="100">
 </a>
 <a href="https://www.bayer.com/">
-    <img src="{% static 'pharmacy/images/partners/bayer.png' %}" alt="Bayer" height="100">
+    <img src="{% static 'info/images/partners/bayer.png' %}" alt="Bayer" height="100">
 </a>
 <a href="https://www.novartis.com/">
-    <img src="{% static 'pharmacy/images/partners/novartis.png' %}" alt="Novartis" height="100">
+    <img src="{% static 'info/images/partners/novartis.png' %}" alt="Novartis" height="100">
 </a>
 <a href="https://www.pfizer.com/">
-    <img src="{% static 'pharmacy/images/partners/pfizer.png' %}" alt="Pfizer" height="100">
+    <img src="{% static 'info/images/partners/pfizer.png' %}" alt="Pfizer" height="100">
 </a>
 
 {% endblock %}
