--- pharmacy/static/pharmacy/css/styles.css
@@ -1,5 +1,0 @@
-.sidebar-nav {
-    margin-top: 20px;
-    padding: 0;
-    list-style: none;
-}
